almost done with mrdsrg-spin-integrated

+++ src/mrdsrg-spin-integrated/dsrg_mrpt2.cc
@@ -2728,7 +2728,7 @@ ambit::Tensor DSRG_MRPT2::separate_tensor(ambit::Tensor& tens, const Dimension& 
                                           const int& h) {
     // test tens and irrep
     int tens_dim = static_cast<int>(tens.dim(0));
-    if (tens_dim != irrep.sum() || tens_dim != tens.dim(1)) {
+    if (tens_dim != irrep.sum() || tens_dim != static_cast<int>(tens.dim(1))) {
         throw PSIEXCEPTION("Wrong dimension for the to-be-separated ambit Tensor.");
     }
     if (h >= irrep.n()) {
